agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "agents using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.100.1,<1.0.0",
    "pymongo>=4.6.1",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.1",
    "pytz>=2024.1",
    "google-genai>=0.1.0"
]

[project.scripts]
agents = "agents.main:run"
run_crew = "agents.main:run"
train = "agents.main:train"
replay = "agents.main:replay"
test = "agents.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
